+++ simple_history/tests/tests/test_utils.py
@@ -1,9 +1,11 @@
 import unittest
+from collections.abc import Callable
 from dataclasses import dataclass
 from datetime import datetime
-from typing import Optional, Type
+from enum import Enum, auto
+from typing import Final, List, Optional, Type, Union
 from unittest import skipUnless
-from unittest.mock import Mock, patch
+from unittest.mock import ANY, Mock, patch
 
 import django
 from django.contrib.auth import get_user_model
@@ -14,16 +16,20 @@ from django.utils import timezone
 
 from simple_history.exceptions import AlternativeManagerError, NotHistoricalModelError
 from simple_history.manager import HistoryManager
-from simple_history.models import HistoricalChanges
+from simple_history.models import HistoricalChanges, HistoricalRecords
 from simple_history.utils import (
+    DisableHistoryInfo,
+    _StoredDisableHistoryInfo,
     bulk_create_with_history,
     bulk_update_with_history,
+    disable_history,
     get_historical_records_of_instance,
     get_history_manager_for_model,
     get_history_model_for_model,
     get_m2m_field_name,
     get_m2m_reverse_field_name,
     get_pk_name,
+    is_history_disabled,
     update_change_reason,
 )
 
@@ -85,8 +91,459 @@ from ..models import (
     TrackedWithConcreteBase,
     Voter,
 )
+from .utils import HistoricalTestCase
 
 User = get_user_model()
+
+
+class DisableHistoryTestCase(HistoricalTestCase):
+    """Tests related to the ``disable_history()`` context manager."""
+
+    def test_disable_history_info(self):
+        """Test that the various utilities for checking the current info on how
+        historical record creation is disabled, return the expected values.
+        This includes ``DisableHistoryInfo`` and ``is_history_disabled()``, as well as
+        the ``_StoredDisableHistoryInfo`` stored through ``HistoricalRecords.context``.
+        """
+
+        class DisableHistoryMode(Enum):
+            NOT_DISABLED = auto()
+            GLOBALLY = auto()
+            PREDICATE = auto()
+
+        poll1 = Poll.objects.create(question="question?", pub_date=timezone.now())
+        poll2 = Poll.objects.create(question="ignore this", pub_date=timezone.now())
+
+        def assert_disable_history_info(
+            mode: DisableHistoryMode, predicate_target: Union[Type[Poll], Poll] = None
+        ):
+            # Check the stored info
+            attr_name = _StoredDisableHistoryInfo.LOCAL_STORAGE_ATTR_NAME
+            info = getattr(HistoricalRecords.context, attr_name, None)
+            if mode is DisableHistoryMode.NOT_DISABLED:
+                self.assertIsNone(info)
+            elif mode is DisableHistoryMode.GLOBALLY:
+                self.assertEqual(
+                    info, _StoredDisableHistoryInfo(instance_predicate=None)
+                )
+            elif mode is DisableHistoryMode.PREDICATE:
+                self.assertEqual(
+                    info, _StoredDisableHistoryInfo(instance_predicate=ANY)
+                )
+                self.assertIsInstance(info.instance_predicate, Callable)
+
+            # Check `DisableHistoryInfo`
+            info = DisableHistoryInfo.get()
+            self.assertEqual(info.not_disabled, mode == DisableHistoryMode.NOT_DISABLED)
+            self.assertEqual(
+                info.disabled_globally, mode == DisableHistoryMode.GLOBALLY
+            )
+            self.assertEqual(
+                info.disabled_for(poll1),
+                predicate_target is Poll or predicate_target == poll1,
+            )
+            self.assertEqual(
+                info.disabled_for(poll2),
+                predicate_target is Poll or predicate_target == poll2,
+            )
+
+            # Check `is_history_disabled()`
+            self.assertEqual(is_history_disabled(), mode == DisableHistoryMode.GLOBALLY)
+            self.assertEqual(
+                is_history_disabled(poll1),
+                mode == DisableHistoryMode.GLOBALLY
+                or predicate_target is Poll
+                or predicate_target == poll1,
+            )
+            self.assertEqual(
+                is_history_disabled(poll2),
+                mode == DisableHistoryMode.GLOBALLY
+                or predicate_target is Poll
+                or predicate_target == poll2,
+            )
+
+        assert_disable_history_info(DisableHistoryMode.NOT_DISABLED)
+
+        with disable_history():
+            assert_disable_history_info(DisableHistoryMode.GLOBALLY)
+
+        assert_disable_history_info(DisableHistoryMode.NOT_DISABLED)
+
+        with disable_history(only_for_model=Poll):
+            assert_disable_history_info(DisableHistoryMode.PREDICATE, Poll)
+
+        assert_disable_history_info(DisableHistoryMode.NOT_DISABLED)
+
+        with disable_history(instance_predicate=lambda poll: "ignore" in poll.question):
+            assert_disable_history_info(DisableHistoryMode.PREDICATE, poll2)
+
+        assert_disable_history_info(DisableHistoryMode.NOT_DISABLED)
+
+    @staticmethod
+    def _test_disable_poll_history(**kwargs):
+        """Create, update and delete some ``Poll`` instances outside and inside
+        the context manager."""
+        last_pk = 0
+
+        def manipulate_poll(
+            poll: Poll = None, *, create=False, update=False, delete=False
+        ) -> Poll:
+            if create:
+                nonlocal last_pk
+                last_pk += 1
+                poll = Poll.objects.create(
+                    pk=last_pk, question=f"qUESTION {last_pk}?", pub_date=timezone.now()
+                )
+            if update:
+                poll.question = f"Question {poll.pk}!"
+                poll.save()
+            if delete:
+                poll.delete()
+            return poll
+
+        poll1 = manipulate_poll(create=True, update=True, delete=True)  # noqa: F841
+        poll2 = manipulate_poll(create=True, update=True)
+        poll3 = manipulate_poll(create=True)
+
+        with disable_history(**kwargs):
+            manipulate_poll(poll2, delete=True)
+            manipulate_poll(poll3, update=True)
+            poll4 = manipulate_poll(create=True, update=True, delete=True)  # noqa: F841
+            poll5 = manipulate_poll(create=True, update=True)
+            poll6 = manipulate_poll(create=True)
+
+        manipulate_poll(poll5, delete=True)
+        manipulate_poll(poll6, update=True)
+        poll7 = manipulate_poll(create=True, update=True, delete=True)  # noqa: F841
+
+    expected_poll_records_before_disable: Final = [
+        {"id": 1, "question": "qUESTION 1?", "history_type": "+"},
+        {"id": 1, "question": "Question 1!", "history_type": "~"},
+        {"id": 1, "question": "Question 1!", "history_type": "-"},
+        {"id": 2, "question": "qUESTION 2?", "history_type": "+"},
+        {"id": 2, "question": "Question 2!", "history_type": "~"},
+        {"id": 3, "question": "qUESTION 3?", "history_type": "+"},
+    ]
+    expected_poll_records_during_disable: Final = [
+        {"id": 2, "question": "Question 2!", "history_type": "-"},
+        {"id": 3, "question": "Question 3!", "history_type": "~"},
+        {"id": 4, "question": "qUESTION 4?", "history_type": "+"},
+        {"id": 4, "question": "Question 4!", "history_type": "~"},
+        {"id": 4, "question": "Question 4!", "history_type": "-"},
+        {"id": 5, "question": "qUESTION 5?", "history_type": "+"},
+        {"id": 5, "question": "Question 5!", "history_type": "~"},
+        {"id": 6, "question": "qUESTION 6?", "history_type": "+"},
+    ]
+    expected_poll_records_after_disable: Final = [
+        {"id": 5, "question": "Question 5!", "history_type": "-"},
+        {"id": 6, "question": "Question 6!", "history_type": "~"},
+        {"id": 7, "question": "qUESTION 7?", "history_type": "+"},
+        {"id": 7, "question": "Question 7!", "history_type": "~"},
+        {"id": 7, "question": "Question 7!", "history_type": "-"},
+    ]
+
+    @staticmethod
+    def _test_disable_poll_with_m2m_history(**kwargs):
+        """Create some ``PollWithManyToMany`` instances and add, remove, set and clear
+        their ``Place`` relations outside and inside the context manager."""
+        last_pk = 0
+        place1 = Place.objects.create(pk=1, name="1")
+        place2 = Place.objects.create(pk=2, name="2")
+
+        def manipulate_places(
+            poll=None, *, add=False, remove=False, set=False, clear=False
+        ) -> PollWithManyToMany:
+            if not poll:
+                nonlocal last_pk
+                last_pk += 1
+                poll = PollWithManyToMany.objects.create(
+                    pk=last_pk, question=f"{last_pk}?", pub_date=timezone.now()
+                )
+            if add:
+                poll.places.add(place1)
+            if remove:
+                poll.places.remove(place1)
+            if set:
+                poll.places.set([place2])
+            if clear:
+                poll.places.clear()
+            return poll
+
+        poll1 = manipulate_places(  # noqa: F841
+            add=True, remove=True, set=True, clear=True
+        )
+        poll2 = manipulate_places(add=True, remove=True, set=True)
+        poll3 = manipulate_places(add=True, remove=True)
+        poll4 = manipulate_places(add=True)
+
+        with disable_history(**kwargs):
+            manipulate_places(poll2, clear=True)
+            manipulate_places(poll3, set=True, clear=True)
+            manipulate_places(poll4, remove=True, set=True, clear=True)
+            poll5 = manipulate_places(  # noqa: F841
+                add=True, remove=True, set=True, clear=True
+            )
+            poll6 = manipulate_places(add=True, remove=True, set=True)
+            poll7 = manipulate_places(add=True, remove=True)
+            poll8 = manipulate_places(add=True)
+
+        manipulate_places(poll6, clear=True)
+        manipulate_places(poll7, set=True, clear=True)
+        manipulate_places(poll8, remove=True, set=True, clear=True)
+        poll9 = manipulate_places(  # noqa: F841
+            add=True, remove=True, set=True, clear=True
+        )
+
+    expected_poll_with_m2m_records_before_disable: Final = [
+        {"id": 1, "question": "1?", "history_type": "+", "places": []},
+        {"id": 1, "question": "1?", "history_type": "~", "places": [Place(pk=1)]},
+        {"id": 1, "question": "1?", "history_type": "~", "places": []},
+        {"id": 1, "question": "1?", "history_type": "~", "places": [Place(pk=2)]},
+        {"id": 1, "question": "1?", "history_type": "~", "places": []},
+        {"id": 2, "question": "2?", "history_type": "+", "places": []},
+        {"id": 2, "question": "2?", "history_type": "~", "places": [Place(pk=1)]},
+        {"id": 2, "question": "2?", "history_type": "~", "places": []},
+        {"id": 2, "question": "2?", "history_type": "~", "places": [Place(pk=2)]},
+        {"id": 3, "question": "3?", "history_type": "+", "places": []},
+        {"id": 3, "question": "3?", "history_type": "~", "places": [Place(pk=1)]},
+        {"id": 3, "question": "3?", "history_type": "~", "places": []},
+        {"id": 4, "question": "4?", "history_type": "+", "places": []},
+        {"id": 4, "question": "4?", "history_type": "~", "places": [Place(pk=1)]},
+    ]
+    expected_poll_with_m2m_records_during_disable: Final = [
+        {"id": 2, "question": "2?", "history_type": "~", "places": []},
+        {"id": 3, "question": "3?", "history_type": "~", "places": [Place(pk=2)]},
+        {"id": 3, "question": "3?", "history_type": "~", "places": []},
+        {"id": 4, "question": "4?", "history_type": "~", "places": []},
+        {"id": 4, "question": "4?", "history_type": "~", "places": [Place(pk=2)]},
+        {"id": 4, "question": "4?", "history_type": "~", "places": []},
+        {"id": 5, "question": "5?", "history_type": "+", "places": []},
+        {"id": 5, "question": "5?", "history_type": "~", "places": [Place(pk=1)]},
+        {"id": 5, "question": "5?", "history_type": "~", "places": []},
+        {"id": 5, "question": "5?", "history_type": "~", "places": [Place(pk=2)]},
+        {"id": 5, "question": "5?", "history_type": "~", "places": []},
+        {"id": 6, "question": "6?", "history_type": "+", "places": []},
+        {"id": 6, "question": "6?", "history_type": "~", "places": [Place(pk=1)]},
+        {"id": 6, "question": "6?", "history_type": "~", "places": []},
+        {"id": 6, "question": "6?", "history_type": "~", "places": [Place(pk=2)]},
+        {"id": 7, "question": "7?", "history_type": "+", "places": []},
+        {"id": 7, "question": "7?", "history_type": "~", "places": [Place(pk=1)]},
+        {"id": 7, "question": "7?", "history_type": "~", "places": []},
+        {"id": 8, "question": "8?", "history_type": "+", "places": []},
+        {"id": 8, "question": "8?", "history_type": "~", "places": [Place(pk=1)]},
+    ]
+    expected_poll_with_m2m_records_after_disable: Final = [
+        {"id": 6, "question": "6?", "history_type": "~", "places": []},
+        {"id": 7, "question": "7?", "history_type": "~", "places": [Place(pk=2)]},
+        {"id": 7, "question": "7?", "history_type": "~", "places": []},
+        {"id": 8, "question": "8?", "history_type": "~", "places": []},
+        {"id": 8, "question": "8?", "history_type": "~", "places": [Place(pk=2)]},
+        {"id": 8, "question": "8?", "history_type": "~", "places": []},
+        {"id": 9, "question": "9?", "history_type": "+", "places": []},
+        {"id": 9, "question": "9?", "history_type": "~", "places": [Place(pk=1)]},
+        {"id": 9, "question": "9?", "history_type": "~", "places": []},
+        {"id": 9, "question": "9?", "history_type": "~", "places": [Place(pk=2)]},
+        {"id": 9, "question": "9?", "history_type": "~", "places": []},
+    ]
+
+    def test__disable_history__with_no_args(self):
+        """Test that no historical records are created inside the context manager with
+        no arguments (i.e. history is globally disabled)."""
+        # Test with `Poll` instances
+        self._test_disable_poll_history()
+        expected_records = [
+            *self.expected_poll_records_before_disable,
+            *self.expected_poll_records_after_disable,
+        ]
+        self.assert_all_records_of_model_equal(Poll, expected_records)
+
+        # Test with `PollWithManyToMany` instances
+        self._test_disable_poll_with_m2m_history()
+        expected_records = [
+            *self.expected_poll_with_m2m_records_before_disable,
+            *self.expected_poll_with_m2m_records_after_disable,
+        ]
+        self.assert_all_records_of_model_equal(PollWithManyToMany, expected_records)
+
+    def test__disable_history__with__only_for_model__poll(self):
+        """Test that no historical records are created for ``Poll`` instances inside
+        the context manager with ``only_for_model=Poll`` as argument."""
+        # Test with `Poll` instances
+        self._test_disable_poll_history(only_for_model=Poll)
+        expected_records = [
+            *self.expected_poll_records_before_disable,
+            *self.expected_poll_records_after_disable,
+        ]
+        self.assert_all_records_of_model_equal(Poll, expected_records)
+
+        # Test with `PollWithManyToMany` instances
+        self._test_disable_poll_with_m2m_history(only_for_model=Poll)
+        expected_records = [
+            *self.expected_poll_with_m2m_records_before_disable,
+            *self.expected_poll_with_m2m_records_during_disable,
+            *self.expected_poll_with_m2m_records_after_disable,
+        ]
+        self.assert_all_records_of_model_equal(PollWithManyToMany, expected_records)
+
+    def test__disable_history__with__only_for_model__poll_with_m2m(self):
+        """Test that no historical records are created for ``PollWithManyToMany``
+        instances inside the context manager with ``only_for_model=PollWithManyToMany``
+        as argument."""
+        # Test with `Poll` instances
+        self._test_disable_poll_history(only_for_model=PollWithManyToMany)
+        expected_records = [
+            *self.expected_poll_records_before_disable,
+            *self.expected_poll_records_during_disable,
+            *self.expected_poll_records_after_disable,
+        ]
+        self.assert_all_records_of_model_equal(Poll, expected_records)
+
+        # Test with `PollWithManyToMany` instances
+        self._test_disable_poll_with_m2m_history(only_for_model=PollWithManyToMany)
+        expected_records = [
+            *self.expected_poll_with_m2m_records_before_disable,
+            *self.expected_poll_with_m2m_records_after_disable,
+        ]
+        self.assert_all_records_of_model_equal(PollWithManyToMany, expected_records)
+
+    def test__disable_history__with__instance_predicate(self):
+        """Test that no historical records are created inside the context manager, for
+        model instances that match the provided ``instance_predicate`` argument."""
+        # Test with `Poll` instances
+        self._test_disable_poll_history(instance_predicate=lambda poll: poll.pk == 4)
+        expected_records = [
+            *self.expected_poll_records_before_disable,
+            *filter(
+                lambda poll_dict: poll_dict["id"] != 4,
+                self.expected_poll_records_during_disable,
+            ),
+            *self.expected_poll_records_after_disable,
+        ]
+        self.assert_all_records_of_model_equal(Poll, expected_records)
+
+        # Test with `PollWithManyToMany` instances
+        self._test_disable_poll_with_m2m_history(
+            instance_predicate=lambda poll: poll.pk == 5
+        )
+        expected_records = [
+            *self.expected_poll_with_m2m_records_before_disable,
+            *filter(
+                lambda poll_dict: poll_dict["id"] != 5,
+                self.expected_poll_with_m2m_records_during_disable,
+            ),
+            *self.expected_poll_with_m2m_records_after_disable,
+        ]
+        self.assert_all_records_of_model_equal(PollWithManyToMany, expected_records)
+
+    def test__disable_history__for_queryset_delete(self):
+        """Test that no historical records are created inside the context manager when
+        deleting objects using the ``delete()`` queryset method."""
+        Poll.objects.create(pk=1, question="delete me", pub_date=timezone.now())
+        Poll.objects.create(pk=2, question="keep me", pub_date=timezone.now())
+        Poll.objects.create(pk=3, question="keep me", pub_date=timezone.now())
+        Poll.objects.create(pk=4, question="delete me", pub_date=timezone.now())
+
+        with disable_history():
+            Poll.objects.filter(question__startswith="delete").delete()
+
+        expected_records = [
+            {"id": 1, "question": "delete me", "history_type": "+"},
+            {"id": 2, "question": "keep me", "history_type": "+"},
+            {"id": 3, "question": "keep me", "history_type": "+"},
+            {"id": 4, "question": "delete me", "history_type": "+"},
+        ]
+        self.assert_all_records_of_model_equal(Poll, expected_records)
+
+        Poll.objects.all().delete()
+        expected_records += [
+            # Django reverses the order before sending the `post_delete` signals
+            # while bulk-deleting
+            {"id": 3, "question": "keep me", "history_type": "-"},
+            {"id": 2, "question": "keep me", "history_type": "-"},
+        ]
+        self.assert_all_records_of_model_equal(Poll, expected_records)
+
+    def test__disable_history__for_foreign_key_cascade_delete(self):
+        """Test that no historical records are created inside the context manager when
+        indirectly deleting objects through a foreign key relationship with
+        ``on_delete=CASCADE``."""
+        poll1 = Poll.objects.create(pk=1, pub_date=timezone.now())
+        poll2 = Poll.objects.create(pk=2, pub_date=timezone.now())
+        Choice.objects.create(pk=11, poll=poll1, votes=0)
+        Choice.objects.create(pk=12, poll=poll1, votes=0)
+        Choice.objects.create(pk=21, poll=poll2, votes=0)
+        Choice.objects.create(pk=22, poll=poll2, votes=0)
+
+        with disable_history():
+            poll1.delete()
+
+        expected_records = [
+            {"id": 11, "poll_id": 1, "history_type": "+"},
+            {"id": 12, "poll_id": 1, "history_type": "+"},
+            {"id": 21, "poll_id": 2, "history_type": "+"},
+            {"id": 22, "poll_id": 2, "history_type": "+"},
+        ]
+        self.assert_all_records_of_model_equal(Choice, expected_records)
+
+        poll2.delete()
+        expected_records += [
+            # Django reverses the order before sending the `post_delete` signals
+            # while bulk-deleting
+            {"id": 22, "poll_id": 2, "history_type": "-"},
+            {"id": 21, "poll_id": 2, "history_type": "-"},
+        ]
+        self.assert_all_records_of_model_equal(Choice, expected_records)
+
+    def assert_all_records_of_model_equal(
+        self, model: Type[Model], expected_records: List[dict]
+    ):
+        records = model.history.all()
+        self.assertEqual(len(records), len(expected_records))
+        for record, expected_record in zip(reversed(records), expected_records):
+            with self.subTest(record=record, expected_record=expected_record):
+                self.assertRecordValues(record, model, expected_record)
+
+    def test_providing_illegal_arguments_fails(self):
+        """Test that providing various illegal arguments and argument combinations
+        fails."""
+
+        def predicate(_):
+            return True
+
+        # Providing both arguments should fail
+        with self.assertRaises(ValueError):
+            with disable_history(only_for_model=Poll, instance_predicate=predicate):
+                pass
+        # Providing the arguments individually should not fail
+        with disable_history(only_for_model=Poll):
+            pass
+        with disable_history(instance_predicate=predicate):
+            pass
+
+        # Passing non-history-tracked models should fail
+        with self.assertRaises(NotHistoricalModelError):
+            with disable_history(only_for_model=Place):
+                pass
+        # Passing non-history-tracked model instances should fail
+        place = Place.objects.create()
+        with self.assertRaises(NotHistoricalModelError):
+            is_history_disabled(place)
+
+    def test_nesting_fails(self):
+        """Test that nesting ``disable_history()`` contexts fails."""
+        # Nesting (twice or more) should fail
+        with self.assertRaises(AssertionError):
+            with disable_history():
+                with disable_history():
+                    pass
+        with self.assertRaises(AssertionError):
+            with disable_history():
+                with disable_history():
+                    with disable_history():
+                        pass
+        # No nesting should not fail
+        with disable_history():
+            pass
 
 
 class UpdateChangeReasonTestCase(TestCase):
@@ -412,8 +869,17 @@ class BulkCreateWithHistoryTestCase(TestCase):
         self.assertEqual(Poll.history.count(), 5)
 
     @override_settings(SIMPLE_HISTORY_ENABLED=False)
-    def test_bulk_create_history_with_disabled_setting(self):
-        bulk_create_with_history(self.data, Poll)
+    def test_bulk_create_history_without_history_enabled(self):
+        with self.assertNumQueries(1):
+            bulk_create_with_history(self.data, Poll)
+
+        self.assertEqual(Poll.objects.count(), 5)
+        self.assertEqual(Poll.history.count(), 0)
+
+    def test_bulk_create_history_with__disable_history(self):
+        with self.assertNumQueries(1):
+            with disable_history(only_for_model=Poll):
+                bulk_create_with_history(self.data, Poll)
 
         self.assertEqual(Poll.objects.count(), 5)
         self.assertEqual(Poll.history.count(), 0)
@@ -667,13 +1133,20 @@ class BulkUpdateWithHistoryTestCase(TestCase):
 
     @override_settings(SIMPLE_HISTORY_ENABLED=False)
     def test_bulk_update_history_without_history_enabled(self):
+        # 5 records from `setUp()`
         self.assertEqual(Poll.history.count(), 5)
-        # because setup called with enabled settings
-        bulk_update_with_history(
-            self.data,
-            Poll,
-            fields=["question"],
-        )
+        bulk_update_with_history(self.data, Poll, fields=["question"])
+
+        self.assertEqual(Poll.objects.count(), 5)
+        self.assertEqual(Poll.objects.get(id=4).question, "Updated question")
+        self.assertEqual(Poll.history.count(), 5)
+        self.assertEqual(Poll.history.filter(history_type="~").count(), 0)
+
+    def test_bulk_update_history_with__disable_history(self):
+        # 5 records from `setUp()`
+        self.assertEqual(Poll.history.count(), 5)
+        with disable_history(only_for_model=Poll):
+            bulk_update_with_history(self.data, Poll, fields=["question"])
 
         self.assertEqual(Poll.objects.count(), 5)
         self.assertEqual(Poll.objects.get(id=4).question, "Updated question")